Add methods of checking new operators

+++ Baram_Project_2022_second_half_GA/NeuralNetwork.h
@@ -6,33 +6,33 @@ private:
 	//전체 가중치 묶음
 	MatrixXd* matrix_weight_array;
 	//전체 가중치 묶음의 개수
-	int weight_matrix_count;
+	size_t weight_matrix_count;
 
 	//입력 계층 노드 개수
-	int input_node_count;
+	size_t input_node_count;
 
 	//은닉 계층 개수
-	int hidden_layer_count;
+	size_t hidden_layer_count;
 	//은닉 계층 노드 각 개수
-	int* each_hidden_node_count;
+	size_t* each_hidden_node_count;
 
 	//출력 계층 노드 개수
-	int output_node_count;
+	size_t output_node_count;
 
-	int* each_matrix_rows;
-	int* each_matrix_cols;
+	size_t* each_matrix_rows;
+	size_t* each_matrix_cols;
 	string ActivationFunction;
 public:
 	NeuralNetwork();
-	NeuralNetwork(const vector<int> each_layer_node_count);
+	NeuralNetwork(const vector<size_t> each_layer_node_count);
 	~NeuralNetwork();
 
 	//신경망을 만드는 함수
-	void make_neural_network(const vector<int> each_layer_node_count);
+	void make_neural_network(const vector<size_t> each_layer_node_count);
 
 	//신경망에 있는 가중치 값을 설정해주는 함수 단, 현재 신경망의 노드의 개수 크기가 같아야 한다.
 	//저장된 값을 불러와서 설정할때 그리고 가중치 값을 업데이트 할때 사용한다.
-	void set_weight(MatrixXd value[], int size);
+	void set_weight(MatrixXd value[], size_t size);
 
 	//모든 가중치 값을 -1과 1사이의 랜덤으로 초기화 할때 사용
 	void all_weight_reset_random();
@@ -50,7 +50,7 @@ public:
 	void PrintMatrix(MatrixXd matrix) const;
 
 	MatrixXd* ReturnAllWeightMatrix();
-	int GetWeightMatrixCount();
+	size_t GetWeightMatrixCount();
 	void SetActivationFunction(string function_name);
 };
